Minor rworking on the abstract class, and fix some of the phpstan errors
